Move frontend completions code to `/completions`
