+++ packages/jupyter-ai/src/completions/index.ts
@@ -1,0 +1,1 @@
+export { inlineCompletionProvider } from './plugin';
