+++ packages/jupyter-ai/src/handler.ts
@@ -2,11 +2,6 @@ import { URLExt } from '@jupyterlab/coreutils';
 
 import { ServerConnection } from '@jupyterlab/services';
 
-import type {
-  IInlineCompletionList,
-  IInlineCompletionItem
-} from '@jupyterlab/completer';
-
 const API_NAMESPACE = 'api/ai';
 
 /**
@@ -60,47 +55,6 @@ export namespace AiService {
     prompt: string;
   };
 
-  export type InlineCompletionRequest = {
-    number: number;
-    path?: string;
-    /* The model has to complete given prefix */
-    prefix: string;
-    /* The model may consider the following suffix */
-    suffix: string;
-    mime: string;
-    /* Whether to stream the response (if streaming is supported by the model) */
-    stream: boolean;
-    language?: string;
-    cell_id?: string;
-  };
-
-  export type CompletionError = {
-    type: string;
-    traceback: string;
-  };
-
-  export type InlineCompletionReply = {
-    /**
-     * Type for this message can be skipped (`inline_completion` is presumed default).
-     **/
-    type?: 'inline_completion';
-    list: IInlineCompletionList;
-    reply_to: number;
-    error?: CompletionError;
-  };
-
-  export type InlineCompletionStreamChunk = {
-    type: 'stream';
-    response: IInlineCompletionItem;
-    reply_to: number;
-    done: boolean;
-  };
-
-  export type InlineCompletionModelChanged = {
-    type: 'model_changed';
-    model: string;
-  };
-
   export type Collaborator = {
     username: string;
     initials: string;
@@ -145,12 +99,6 @@ export namespace AiService {
     | HumanChatMessage
     | ConnectionMessage
     | ClearMessage;
-
-  export type CompleterMessage =
-    | InlineCompletionReply
-    | ConnectionMessage
-    | InlineCompletionStreamChunk
-    | InlineCompletionModelChanged;
 
   export type ChatHistory = {
     messages: ChatMessage[];
