+++ packages/jupyter-ai/src/index.ts
@@ -12,7 +12,7 @@ import { buildChatSidebar } from './widgets/chat-sidebar';
 import { SelectionWatcher } from './selection-watcher';
 import { ChatHandler } from './chat_handler';
 import { buildErrorWidget } from './widgets/chat-error';
-import { inlineCompletionProvider } from './inline-completions';
+import { inlineCompletionProvider } from './completions';
 import { jupyternautStatus } from './status';
 
 export type DocumentTracker = IWidgetTracker<IDocumentWidget>;
